# regalloc: eliminates unnecessary Instr loop