This eliminates the unnecessary Instructions loop in
register allocation by sacrificing an almost negligible 
memory space. As a result,  overall compilation gets
1% faster than before.

```
goos: darwin
goarch: arm64
pkg: github.com/tetratelabs/wazero
                      │  old.txt   │             new.txt              │
                      │   sec/op   │   sec/op    vs base              │
Compilation/wazero-10   1.973 ± 1%   1.945 ± 0%  -1.42% (p=0.002 n=6)
Compilation/zig-10      4.128 ± 0%   4.068 ± 1%  -1.45% (p=0.002 n=6)
Compilation/zz-10       18.32 ± 0%   18.08 ± 1%  -1.27% (p=0.002 n=6)
geomean                 5.303        5.230       -1.38%

                      │   old.txt    │              new.txt               │
                      │     B/op     │     B/op      vs base              │
Compilation/wazero-10   284.4Mi ± 0%   284.8Mi ± 0%  +0.13% (p=0.002 n=6)
Compilation/zig-10      590.1Mi ± 0%   593.3Mi ± 0%  +0.55% (p=0.002 n=6)
Compilation/zz-10       536.4Mi ± 0%   536.8Mi ± 0%  +0.06% (p=0.002 n=6)
geomean                 448.2Mi        449.3Mi       +0.25%

                      │   old.txt   │              new.txt              │
                      │  allocs/op  │  allocs/op   vs base              │
Compilation/wazero-10   429.1k ± 0%   429.1k ± 0%       ~ (p=0.937 n=6)
Compilation/zig-10      269.7k ± 0%   269.6k ± 0%       ~ (p=1.000 n=6)
Compilation/zz-10       605.7k ± 0%   606.0k ± 0%       ~ (p=0.132 n=6)
geomean                 412.3k        412.4k       +0.02%
```

#2182